+++ examples/sedk-postgres-js-example/index.js
@@ -1,16 +1,22 @@
-const importFunctionsList = [
-	async () => await import('./readme-example.js'),
-	async () => await import('./example1.js'),
+const examples = [
+	'./readme-example.js',
+	'./example1.js',
 ]
 
 async function main() {
-	for (const fn of importFunctionsList) {
-		console.log(`📦 run ${fn.name}`)
-		await fn()
+	const functions = examples.map((filename) => ({
+		filename,
+		execFile: async () => await import(filename)
+	}))
+	for (const fn of functions) {
+		console.log(`📦 run ${fn.filename}`)
+		await fn.execFile()
+		console.log('')
 	}
 }
 
 console.log('🚀 Starting sedk-postgres-js-example ...')
+console.log('')
 main().then(() => {
 	console.log('👋 End of sedk-postgres-js-example')
 })